% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAM3ASest.R
\name{CAM3ASest}
\alias{CAM3ASest}
\title{A and S matrix estimation by CAM3}
\usage{
CAM3ASest(MGResult, PrepResult, data, corner.strategy = 2, generalNMF = FALSE)
}
\arguments{
\item{MGResult}{An object of class "\code{\link[debCAM]{CAMMGObj}}" obtained
from \code{\link{CAM3MGCluster}} function.}

\item{PrepResult}{An object of class "\code{\link[debCAM]{CAMPrepObj}}"
obtained from \code{\link[debCAM]{CAMPrep}} function.}

\item{data}{Matrix of mixture expression profiles which need to be the same
as the input of \code{\link[debCAM]{CAMPrep}}.
Data frame, SummarizedExperiment or ExpressionSet object will be
internally coerced into a matrix.
Each row is a gene and each column is a sample.
Data should be in non-log linear space with non-negative numerical values
(i.e. >= 0). Missing values are not supported.
All-zero rows will be removed internally.}

\item{corner.strategy}{The method to detect corner clusters.
1: minimum sum of margin-of-errors; 2: minimum sum of reconstruction
errors. The default is 2.}

\item{generalNMF}{If TRUE, the decomposed proportion matrix has no sum-to-one
constraint for each row. Without this constraint, the scale ambiguity of
each column vector in proportion matrix will not be removed.
The default is FALSE.}
}
\value{
An object of class "\code{\link[debCAM]{CAMASObj}}" containing the
following components:
\item{Aest}{Estimated proportion matrix.}
\item{Sest}{Estimated subpopulation-specific expression matrix.}
\item{Aest.proj}{Estimated proportion matrix, before removing scale
ambiguity.}
\item{Ascale}{The estimated scales to remove scale ambiguity
of each column vector in Aest. Sum-to-one constraint on each row of
Aest is used for scale estimation.}
\item{AestO}{No use in CAM3, will be deprecated in future version.}
\item{SestO}{No use in CAM3, will be deprecated in future version.}
\item{AestO.proj}{No use in CAM3, will be deprecated in future version.}
\item{AscaleO}{No use in CAM3, will be deprecated in future version.}
\item{datalength}{Code length of data.}
\item{mdl}{The mdl value.}
}
\description{
This function estimates A and S matrix based on marker gene clusters
detected by CAM3.
}
\details{
This function is used internally by \code{\link{CAM3Run}} function to
estimate proportion matrix (A), subpopulation-specific expression matrix (S)
and mdl values. It can also be used when you want to perform CAM step by
step.

The mdl value is based on original data with A matrix estimated by
transforming dimension-reduced A matrix back to original space.
The mdl value is the sum of two terms: code length of data under the model
and code length of model. Both mdl value and the first term (code length
of data) will be returned.
}
\examples{
#obtain data
data(ratMix3)
data <- ratMix3$X

#preprocess data
rPrep3 <- CAMPrep3(data, dim.rdc = 3, thres.low = 0.30, thres.high = 0.95)

#Marker gene cluster detection with greedy search
rMGC3 <- CAM3MGCluster(3, rPrep3)

#A and S matrix estimation
rASest3 <- CAM3ASest(rMGC3, rPrep3, data)
}
