% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAM3MGCluster.R
\name{CAM3MGCluster}
\alias{CAM3MGCluster}
\title{MG cluster detection for CAM3}
\usage{
CAM3MGCluster(PrepResult, fast.mode = TRUE)
}
\arguments{
\item{PrepResult}{An object of class "\code{\link[debCAM]{CAMPrepObj}}"
obtained from \code{\link[debCAM]{CAMPrep}} function.}

\item{fast.mode}{Use fast mode of greedy search or not. The normal mode may
give more accurate results, but computation time is much longer. The
default is TRUE.}
}
\value{
A list of class "\code{\link[debCAM]{CAMMGObj}}" containing the
following components:
\item{idx}{No use in CAM3, will be deprecated in future version.}
\item{corner}{The indexes of clusters as detected corners. The first row is
detected by the forward greedy search, and the second row is detected by
the backward greedy search.}
\item{error}{Two values. The first one is the reconstruction error of forward
greedy search, and the second one is for the backward greedy search.}
}
\description{
This function finds corner clusters as MG clusters
(clusters containing marker genes).
}
\details{
This function is used internally by \code{\link{CAM3Run}} function
to detect clusters containing marker genes,
or used when you want to perform CAM step by step.

This function provides two solutions. The forward and backward greedy search.
Then the best one of certain source number is selected based on
reconstruction errors of all data points in original space.
}
\examples{
#obtain data
data(ratMix3)
data <- ratMix3$X

#preprocess data
rPrep3 <- CAM3Prep(data, dim.rdc = 3, thres.low = 0.30, thres.high = 0.95)

#Marker gene cluster detection with greedy search
rMGC3 <- CAM3MGCluster(rPrep)
}
