% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotMG.R
\name{cotMG}
\alias{cotMG}
\title{Select markers in S matrix with COT}
\usage{
cotMG(
  data = NULL,
  Sest,
  thres.low = 0.05,
  thres.high = 1,
  cos.thres = 1,
  top = NULL,
  per = NULL
)
}
\arguments{
\item{data}{A data set that will be internally coerced into a matrix.
If data is not NULL, "thres.low" and "thres.high" would decide which
genes should be removed in data. If data is NULL, the thresholds work
on "Sest" directly.}

\item{Sest}{Subpopulation-specific expression matrix. Each row is a gene
and each column is a sample.}

\item{thres.low}{The lower bound of percentage of genes to keep with ranked
norm. The value should be between 0 and 1. The default is 0.05.}

\item{thres.high}{The higher bound of percentage of genes to keep with ranked
norm. The value should be between 0 and 1. The default is 1.}

\item{cos.thres}{The cosine threshold for markers.}

\item{top}{The upper bound of total markers.}

\item{per}{The upper bound of markers of each Subpopulation.}
}
\value{
A list of vectors, each of which contains marker genes for one
subpopulation.
}
\description{
This function computes cosine values to select markers.
}
\details{
This function uses COT to detect markers.
}
\examples{
data(ratMix3)
data <- ratMix3$X

#Find marker genes
MGlist <- cotMG(data, rASest3$S, cos.thres = 0.99)
}
